Rename `lib/grib_functions.js` to `lib/cosmo-d2_functions.js`
... in order to use consistent file names.
