+++ lib/cosmo-d2_functions.js
