+++ lib/cosmo-d2_utils.js
@@ -6,7 +6,7 @@
 const path = require('path')
 const moment = require('moment')
 
-const gf = require('./grib_functions')
+const cd2f = require('./cosmo-d2_functions')
 const su = require('./stations_utils')
 const log = require('./logger')
 
@@ -67,7 +67,7 @@ async function loadCosmoD2Forecast (
 ) {
   let timeseriesData
   if (voiConfig.cosmo.functionType === 'loadBaseValue') {
-    timeseriesData = await gf.loadBaseValue({
+    timeseriesData = await cd2f.loadBaseValue({
       gribBaseDirectory,
       referenceTimestamp,
       voi: voiConfig.cosmo.options.key,
@@ -79,7 +79,7 @@ async function loadCosmoD2Forecast (
       targetUnit: voiConfig.target.unit
     })
   } else if (voiConfig.cosmo.functionType === 'loadRadiationValue') {
-    timeseriesData = await gf.loadRadiationValue({
+    timeseriesData = await cd2f.loadRadiationValue({
       gribBaseDirectory,
       referenceTimestamp,
       voi: voiConfig.cosmo.options.key,
@@ -91,7 +91,7 @@ async function loadCosmoD2Forecast (
       targetUnit: voiConfig.target.unit
     })
   } else if (voiConfig.cosmo.functionType === 'load2DVectorNorm') {
-    timeseriesData = await gf.load2DVectorNorm({
+    timeseriesData = await cd2f.load2DVectorNorm({
       gribBaseDirectory,
       referenceTimestamp,
       voi1: voiConfig.cosmo.options.voi1_key,
@@ -105,7 +105,7 @@ async function loadCosmoD2Forecast (
       targetUnit: voiConfig.target.unit
     })
   } else if (voiConfig.cosmo.functionType === 'load2DVectorAngle') {
-    timeseriesData = await gf.load2DVectorAngle({
+    timeseriesData = await cd2f.load2DVectorAngle({
       gribBaseDirectory,
       referenceTimestamp,
       voi1: voiConfig.cosmo.options.voi1_key,
